Make the Keyboard Backlight work again.
I think previously this was for the thinklight? 2 bits, shmoo bits, the driver wants 0x200 and the backlight level.  Unfortunately this doesn't want to work on windows and I'm not sure why yet. Maybe something doesn't load or I need to throw the DLL in IDA and find out what is up. I have not worked out events yet.. perhaps windows operates on those more.